Initial celery implementation
